# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "This is the API documentation for my Laravel application.",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        },
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/resources": {
            "get": {
                "tags": [
                    "Resource"
                ],
                "summary": "Get a list of resources",
                "operationId": "b68ca62605727ce3ea6d4551e462e567",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Resource"
                ],
                "summary": "Post a new resource",
                "operationId": "35423e34cd5fd4cdcd4ec832e2986fa1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Example Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "example@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/resources/{id}": {
            "put": {
                "tags": [
                    "Resource"
                ],
                "summary": "Update an existing resource",
                "operationId": "c20e8736e630a63703369926f0b02866",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "updated@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Resource"
                ],
                "summary": "Delete an existing resource",
                "operationId": "d3169e3c0fab73499b5dcf41024e22bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Resource": {
                "title": "Resource",
                "description": "Resource model schema",
                "required": [
                    "name",
                    "type"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the resource",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the resource",
                        "type": "string",
                        "example": "Sample Resource"
                    },
                    "type": {
                        "description": "Type of the resource",
                        "type": "string",
                        "example": "example_type"
                    },
                    "description": {
                        "description": "Description of the resource",
                        "type": "string",
                        "example": "This is a sample resource description."
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-10T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-10T11:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Resource",
            "description": "Resource"
        }
    ]
}